+++ include/smooth/backends/win32/backendwin32.h
@@ -1,5 +1,5 @@
  /* The smooth Class Library
-  * Copyright (C) 1998-2009 Robert Kausch <robert.kausch@gmx.net>
+  * Copyright (C) 1998-2010 Robert Kausch <robert.kausch@gmx.net>
   *
   * This library is free software; you can redistribute it and/or
   * modify it under the terms of "The Artistic License, Version 2.0".
@@ -37,7 +37,7 @@ namespace smooth
 
 	namespace Backends
 	{
-		const Int	 BACKEND_WIN32 = 2;
+		const Short	 BACKEND_WIN32 = 2;
 
 		class BackendWin32 : public Backend
 		{
