+++ include/smooth/graphics/backends/gdi/fontgdi.h
@@ -1,5 +1,5 @@
  /* The smooth Class Library
-  * Copyright (C) 1998-2009 Robert Kausch <robert.kausch@gmx.net>
+  * Copyright (C) 1998-2010 Robert Kausch <robert.kausch@gmx.net>
   *
   * This library is free software; you can redistribute it and/or
   * modify it under the terms of "The Artistic License, Version 2.0".
@@ -25,12 +25,12 @@ namespace smooth
 {
 	namespace GUI
 	{
-		const Int	 FONT_GDI	= 1;
+		const Short	 FONT_GDI	= 1;
 
 		class FontGDI : public FontBackend
 		{
 			public:
-					 FontGDI(const String &, Int, Int, Int, const Color &);
+					 FontGDI(const String &, Short, Short, Short, const Color &);
 					~FontGDI();
 
 				Size	 GetTextSize(const String &) const;
