+++ include/smooth/graphics/backends/gdiplus/surfacegdiplus.h
@@ -28,12 +28,12 @@ namespace smooth
 {
 	namespace GUI
 	{
-		const Int	 SURFACE_GDIPLUS	= 4;
+		const Short	 SURFACE_GDIPLUS	= 4;
 
 		class SurfaceGDIPlus : public SurfaceBackend
 		{
 			private:
-				static Int		 surfaceDPI;
+				static Short		 surfaceDPI;
 			protected:
 				HWND			 window;
 
@@ -56,7 +56,7 @@ namespace smooth
 
 				Void			*GetSystemSurface() const;
 
-				Int			 GetSurfaceDPI() const;
+				Short			 GetSurfaceDPI() const;
 
 				Int			 SetPixel(const Point &, const Color &);
