+++ include/smooth/gui/dialogs/tipodaydlg.h
@@ -1,5 +1,5 @@
  /* The smooth Class Library
-  * Copyright (C) 1998-2009 Robert Kausch <robert.kausch@gmx.net>
+  * Copyright (C) 1998-2010 Robert Kausch <robert.kausch@gmx.net>
   *
   * This library is free software; you can redistribute it and/or
   * modify it under the terms of "The Artistic License, Version 2.0".
@@ -75,8 +75,8 @@ namespace smooth
 					Void		 Paint();
 			};
 
-			const Int	 TIP_ORDERED	= 0;
-			const Int	 TIP_RANDOM	= 1;
+			const Short	 TIP_ORDERED	= 0;
+			const Short	 TIP_RANDOM	= 1;
 		};
 	};
 };
