+++ include/smooth/gui/mdi/window.h
@@ -1,5 +1,5 @@
  /* The smooth Class Library
-  * Copyright (C) 1998-2009 Robert Kausch <robert.kausch@gmx.net>
+  * Copyright (C) 1998-2010 Robert Kausch <robert.kausch@gmx.net>
   *
   * This library is free software; you can redistribute it and/or
   * modify it under the terms of "The Artistic License, Version 2.0".
@@ -33,12 +33,12 @@ namespace smooth
 			class SMOOTHAPI Window : public GUI::Window
 			{
 				public:
-					static const Int	 classID;
+					static const Short	 classID;
 
 								 Window(const String &, const Point &, const Size &);
 					virtual			~Window();
 
-					virtual Bool		 IsTypeCompatible(Int) const;
+					virtual Bool		 IsTypeCompatible(Short) const;
 			};
 		};
 	};
