+++ include/smooth/gui/widgets/basic/arrows.h
@@ -1,5 +1,5 @@
  /* The smooth Class Library
-  * Copyright (C) 1998-2009 Robert Kausch <robert.kausch@gmx.net>
+  * Copyright (C) 1998-2010 Robert Kausch <robert.kausch@gmx.net>
   *
   * This library is free software; you can redistribute it and/or
   * modify it under the terms of "The Artistic License, Version 2.0".
@@ -50,7 +50,7 @@ namespace smooth
 				Hotspot			*arrow1Hotspot;
 				Hotspot			*arrow2Hotspot;
 			public:
-				static const Int	 classID;
+				static const Short	 classID;
 
 							 Arrows(const Point &, const Size &, Int = OR_HORZ, Int * = NIL, Int = 0, Int = 100);
 				virtual			~Arrows();
