+++ include/smooth/gui/widgets/basic/button.h
@@ -1,5 +1,5 @@
  /* The smooth Class Library
-  * Copyright (C) 1998-2009 Robert Kausch <robert.kausch@gmx.net>
+  * Copyright (C) 1998-2010 Robert Kausch <robert.kausch@gmx.net>
   *
   * This library is free software; you can redistribute it and/or
   * modify it under the terms of "The Artistic License, Version 2.0".
@@ -27,8 +27,8 @@ namespace smooth
 {
 	namespace GUI
 	{
-		const Int	 BF_NORMAL	= 0;
-		const Int	 BF_NOFRAME	= 1;
+		const Short	 BF_NORMAL	= 0;
+		const Short	 BF_NOFRAME	= 1;
 
 		class SMOOTHAPI Button : public Widget
 		{
@@ -37,7 +37,7 @@ namespace smooth
 
 				Hotspot			*hotspot;
 			public:
-				static const Int	 classID;
+				static const Short	 classID;
 
 							 Button(const String &, const Bitmap &, const Point &, const Size &);
 				virtual			~Button();
