+++ include/smooth/gui/widgets/basic/hyperlink.h
@@ -1,5 +1,5 @@
  /* The smooth Class Library
-  * Copyright (C) 1998-2009 Robert Kausch <robert.kausch@gmx.net>
+  * Copyright (C) 1998-2010 Robert Kausch <robert.kausch@gmx.net>
   *
   * This library is free software; you can redistribute it and/or
   * modify it under the terms of "The Artistic License, Version 2.0".
@@ -35,7 +35,7 @@ namespace smooth
 
 				Hotspot			*hotspot;
 			public:
-				static const Int	 classID;
+				static const Short	 classID;
 
 							 Hyperlink(const String &, const Bitmap &, const String &, const Point &, const Size & = Size(0, 0));
 				virtual			~Hyperlink();
