+++ include/smooth/gui/widgets/basic/progressbar.h
@@ -1,5 +1,5 @@
  /* The smooth Class Library
-  * Copyright (C) 1998-2009 Robert Kausch <robert.kausch@gmx.net>
+  * Copyright (C) 1998-2010 Robert Kausch <robert.kausch@gmx.net>
   *
   * This library is free software; you can redistribute it and/or
   * modify it under the terms of "The Artistic License, Version 2.0".
@@ -26,9 +26,9 @@ namespace smooth
 {
 	namespace GUI
 	{
-		const Int PB_NOTEXT	= 0;
-		const Int PB_VALUE	= 1;
-		const Int PB_PERCENT	= 2;
+		const Short	 PB_NOTEXT	= 0;
+		const Short	 PB_VALUE	= 1;
+		const Short	 PB_PERCENT	= 2;
 
 		class SMOOTHAPI Progressbar : public Widget
 		{
@@ -42,7 +42,7 @@ namespace smooth
 				Int			 startValue;
 				Int			 endValue;
 			public:
-				static const Int	 classID;
+				static const Short	 classID;
 
 							 Progressbar(const Point &, const Size &, Int = OR_HORZ, Int = PB_PERCENT, Int = 0, Int = 100, Int = 0);
 				virtual			~Progressbar();
