+++ include/smooth/gui/widgets/multi/image/imageentry.h
@@ -1,5 +1,5 @@
  /* The smooth Class Library
-  * Copyright (C) 1998-2009 Robert Kausch <robert.kausch@gmx.net>
+  * Copyright (C) 1998-2010 Robert Kausch <robert.kausch@gmx.net>
   *
   * This library is free software; you can redistribute it and/or
   * modify it under the terms of "The Artistic License, Version 2.0".
@@ -33,14 +33,14 @@ namespace smooth
 			protected:
 				Image			*image;
 			public:
-				static const Int	 classID;
+				static const Short	 classID;
 
 							 ImageEntry(const Bitmap &, const Size &);
 				virtual			~ImageEntry();
 
 				virtual Int		 Paint(Int);
 
-				virtual Bool		 IsTypeCompatible(Int) const;
+				virtual Bool		 IsTypeCompatible(Short) const;
 
 				virtual String		 ToString() const				{ return "an ImageEntry"; }
 			slots:
