+++ include/smooth/gui/widgets/multi/list/listbox.h
@@ -1,5 +1,5 @@
  /* The smooth Class Library
-  * Copyright (C) 1998-2009 Robert Kausch <robert.kausch@gmx.net>
+  * Copyright (C) 1998-2010 Robert Kausch <robert.kausch@gmx.net>
   *
   * This library is free software; you can redistribute it and/or
   * modify it under the terms of "The Artistic License, Version 2.0".
@@ -27,12 +27,12 @@ namespace smooth
 {
 	namespace GUI
 	{
-		const Int	 LF_NORMAL		= 0;
-		const Int	 LF_HIDESCROLLBAR	= 1;
-		const Int	 LF_ALLOWRESELECT	= 2;
-		const Int	 LF_MULTICHECKBOX	= 4;
-		const Int	 LF_HIDEHEADER		= 8;
-		const Int	 LF_ALLOWREORDER	= 16;
+		const Short	 LF_NORMAL		= 0;
+		const Short	 LF_HIDESCROLLBAR	= 1;
+		const Short	 LF_ALLOWRESELECT	= 2;
+		const Short	 LF_MULTICHECKBOX	= 4;
+		const Short	 LF_HIDEHEADER		= 8;
+		const Short	 LF_ALLOWREORDER	= 16;
 
 		class SMOOTHAPI ListBox : public List
 		{
@@ -46,7 +46,7 @@ namespace smooth
 
 				Int			 GetEntriesHeight() const;
 			public:
-				static const Int	 classID;
+				static const Short	 classID;
 
 							 ListBox(const Point &, const Size &);
 				virtual			~ListBox();
