+++ include/smooth/gui/widgets/multi/list/listentryseparator.h
@@ -1,5 +1,5 @@
  /* The smooth Class Library
-  * Copyright (C) 1998-2009 Robert Kausch <robert.kausch@gmx.net>
+  * Copyright (C) 1998-2010 Robert Kausch <robert.kausch@gmx.net>
   *
   * This library is free software; you can redistribute it and/or
   * modify it under the terms of "The Artistic License, Version 2.0".
@@ -28,7 +28,7 @@ namespace smooth
 		class SMOOTHAPI ListEntrySeparator : public ListEntry
 		{
 			public:
-				static const Int	 classID;
+				static const Short	 classID;
 
 							 ListEntrySeparator();
 				virtual			~ListEntrySeparator();
@@ -37,7 +37,7 @@ namespace smooth
 
 				virtual String		 ToString() const	{ return "a ListEntrySeparator"; }
 
-				virtual Bool		 IsTypeCompatible(Int) const;
+				virtual Bool		 IsTypeCompatible(Short) const;
 		};
 	};
 };
