+++ include/smooth/gui/widgets/multi/menu/menubar.h
@@ -1,5 +1,5 @@
  /* The smooth Class Library
-  * Copyright (C) 1998-2009 Robert Kausch <robert.kausch@gmx.net>
+  * Copyright (C) 1998-2010 Robert Kausch <robert.kausch@gmx.net>
   *
   * This library is free software; you can redistribute it and/or
   * modify it under the terms of "The Artistic License, Version 2.0".
@@ -25,14 +25,14 @@ namespace smooth
 {
 	namespace GUI
 	{
-		const Int	 MB_GRAYSCALE	= 0;
-		const Int	 MB_COLOR	= 1;
-		const Int	 MB_POPUPOPEN	= 2;
+		const Short	 MB_GRAYSCALE	= 0;
+		const Short	 MB_COLOR	= 1;
+		const Short	 MB_POPUPOPEN	= 2;
 
 		class SMOOTHAPI Menubar : public Menu
 		{
 			public:
-				static const Int	 classID;
+				static const Short	 classID;
 
 							 Menubar();
 				virtual			~Menubar();
