+++ include/smooth/gui/widgets/multi/menu/menubarentry.h
@@ -1,5 +1,5 @@
  /* The smooth Class Library
-  * Copyright (C) 1998-2009 Robert Kausch <robert.kausch@gmx.net>
+  * Copyright (C) 1998-2010 Robert Kausch <robert.kausch@gmx.net>
   *
   * This library is free software; you can redistribute it and/or
   * modify it under the terms of "The Artistic License, Version 2.0".
@@ -33,14 +33,14 @@ namespace smooth
 				Hotspot			*hotspot;
 				Hotspot			*actionHotspot;
 			public:
-				static const Int	 classID;
+				static const Short	 classID;
 
 							 MenubarEntry(const String & = NIL, const Bitmap & = NIL, PopupMenu * = NIL, Bool * = NIL, Int * = NIL, Int = 0);
 				virtual			~MenubarEntry();
 
 				virtual Int		 Paint(Int);
 
-				virtual Bool		 IsTypeCompatible(Int) const;
+				virtual Bool		 IsTypeCompatible(Short) const;
 			slots:
 				Void			 OnMouseOver();
