+++ include/smooth/gui/widgets/multi/menu/popupmenuentrycheck.h
@@ -1,5 +1,5 @@
  /* The smooth Class Library
-  * Copyright (C) 1998-2009 Robert Kausch <robert.kausch@gmx.net>
+  * Copyright (C) 1998-2010 Robert Kausch <robert.kausch@gmx.net>
   *
   * This library is free software; you can redistribute it and/or
   * modify it under the terms of "The Artistic License, Version 2.0".
@@ -28,14 +28,14 @@ namespace smooth
 		class SMOOTHAPI PopupMenuEntryCheck : public PopupMenuEntry
 		{
 			public:
-				static const Int	 classID;
+				static const Short	 classID;
 
 							 PopupMenuEntryCheck(const String &, Bool *);
 				virtual			~PopupMenuEntryCheck();
 
 				virtual Int		 Paint(Int);
 
-				virtual Bool		 IsTypeCompatible(Int) const;
+				virtual Bool		 IsTypeCompatible(Short) const;
 			slots:
 				Void			 OnClickEntry();
 		};
