+++ include/smooth/gui/widgets/special/dragcontrol.h
@@ -1,5 +1,5 @@
  /* The smooth Class Library
-  * Copyright (C) 1998-2009 Robert Kausch <robert.kausch@gmx.net>
+  * Copyright (C) 1998-2010 Robert Kausch <robert.kausch@gmx.net>
   *
   * This library is free software; you can redistribute it and/or
   * modify it under the terms of "The Artistic License, Version 2.0".
@@ -33,7 +33,7 @@ namespace smooth
 			protected:
 				Hotspot			*dragHotspot;
 			public:
-				static const Int	 classID;
+				static const Short	 classID;
 
 							 DragControl();
 				virtual			~DragControl();
